extraction/requirements.txt
openai-whisper
torch
numpy
pandas
pillow
langchain-ollama
ollama
